move slint task to own rust module
