+++ helixflow/src/lib.rs
@@ -4,7 +4,7 @@ use log::debug;
 use slint::ComponentHandle;
 
 use helixflow_slint::HelixFlow;
-use helixflow_slint::blocking::create_task;
+use helixflow_slint::task::blocking::create_task;
 use helixflow_surreal::blocking::SurrealDb;
 
 pub fn run_helixflow() {
